requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cssselect==1.3.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.19.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
hf-xet==1.1.10
httplib2==0.31.0
huggingface-hub==0.35.3
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
lxml_html_clean==0.4.2
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.6.0
networkx==3.5
newspaper3k==0.2.8
nltk==3.9.2
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-file==2.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
sentencepiece==0.2.1
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.22.1
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
validators==0.35.0
